Beta Version
Needs more thorough testing, but no known bugs/problems.

-Fixed the issue with the logger items not being logged properly
-Added a log file tester to confirm readability of log files
-Added raw system time to log files as first element (in game time
obviously changes relative to your time travel and isn't a good index of
total time)
-Fixed a previously unknown bug found during testing in the boundary
color changes that required a rewrite of that module. There was an edge
case where you could reverse time during the transition between
boundaries and trick the system into not changing the color when it
should've. The only fix for this was to move to a profile system where
the state at very short time intervals is always known (stored, not
computed), and the state index is tracked by game
time/recalled/displayed. It's a bit less efficient, but it's reliable.