Item Colors during Events and Invis Bubble
-Reworked items slightly during practice and study so that during their
event, they change texture to a colored version of the same texture
(this does not occur during test)
-Added keyboard option for invisibility to toggle on and off. If present
in the config, it is available. However, by default it is off. This
should only be used in practice to illustrate the invisibility.
-Added a persistent "bubble" glow around the player to see the
invisibility distance. There are hidden keyboard/controller key options
in the config which can be used to toggle it during play (just in case).
The default state is on.
-Fixed a bug in the test logger where items could become reordered when
placed/removed. Also a bug which caused errors preventing logging.
-Updated gitignore to exclude Library, Temp and Obj